api/pyproject.toml
[project]
name = "crm"
version = "0.1.0"
description = "FastAPI, GCP, PostgreSQL, Tortoise"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aerich>=0.8.0",
    "asyncpg>=0.31.0",
    "boto3>=1.35.0",
    "fastapi>=0.124.4",
    "fastapi-users>=15.0.1",
    "fastapi-users-db-tortoise>=2.0.0",
    "gcloud>=0.18.3",
    "google-auth>=2.43.0",
    "gender-guesser>=0.4.0",
    "openai>=1.52.0",
    "python-dotenv>=1.2.1",
    "tortoise-orm>=0.18.1",
    "uvicorn[standard]>=0.38.0",
]

[tool.aerich]
tortoise_orm = "config.local.TORTOISE_ORM"
location = "./migrations"
src_folder = "./."
